pyproject.toml

[tool.poetry]
name = "zeta-software"
version = "0.1.0"
description = "Software for project zeta"
authors = []

[tool.poetry.dependencies]
python = "^3.8"
click = "^8.1.2"
numpy = "1.21.0"
pyserial = "^3.5"

[tool.poetry.dev-dependencies]
pytest = "^7.0.1"
jupyter = "^1.0.0"
snakeviz = "^2.1.1"
yapf = "^0.32.0"
autoflake = "^1.4"
pyproject-autoflake = "^1.0.1"
isort = "^5.10.1"
flake8 = "^4.0.1"
pyproject-flake8 = "^0.0.1-alpha.2"
pre-commit = "^2.17.0"

[tool.autoflake]
in-place = true
remove-all-unused-imports = true
remove-unused-variables = true

[tool.isort]
py_version = 38
profile = "black"
line_length = 120

[tool.flake8]
max-line-length = 120
max-complexity = 10

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
